requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
attrs==19.3.0             # via pytest
click==7.1.2              # via code-annotations
code-annotations==0.3.4   # via -r requirements/test.in
coverage==5.2             # via pytest-cov
importlib-metadata==1.7.0  # via pluggy, pytest, stevedore
jinja2==2.11.2            # via code-annotations
markupsafe==1.1.1         # via jinja2
more-itertools==5.0.0     # via -c requirements/constraints.txt, pytest
packaging==20.4           # via pytest
pbr==5.4.5                # via stevedore
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/test.in
pytest-django==3.9.0      # via -r requirements/test.in
pytest==5.4.3             # via pytest-cov, pytest-django
python-slugify==4.0.1     # via code-annotations
pytz==2020.1              # via -r requirements/base.txt, django
pyyaml==5.3.1             # via code-annotations
six==1.15.0               # via more-itertools, packaging
sqlparse==0.3.1           # via -r requirements/base.txt, django
stevedore==3.1.0          # via code-annotations
text-unidecode==1.3       # via python-slugify
wcwidth==0.2.5            # via pytest
zipp==3.1.0               # via importlib-metadata
